Develop

+++ server.js
@@ -1,9 +1,13 @@
+require("dotenv").config({
+  path: `.env.${process.env.NODE_ENV || "development"}`
+});
+
 const app = require("./src/app");
 const db = require("./src/loaders/models.loader");
 
 const PORT = process.env.PORT || 3000;
 
-if (process.env.DB_SYNC === "true") {
+if (process.env.NODE_ENV !== "production" && process.env.DB_SYNC === "true") {
   db.sequelize
     .sync({ force: false })
     .then(() => {
